+++ app/scripts/kwic_download.js
@@ -1,5 +1,6 @@
 /** @format */
 import _ from "lodash"
+import moment from "moment"
 import CSV from "comma-separated-values/csv"
 import { locObj } from "@/i18n"